Updated missing paramenters for some entities and DTOs.

+++ src/main/java/com/koleff/kare/exercise/models/dto/ExerciseSetDto.java
@@ -5,8 +5,8 @@ import java.util.UUID;
 public record ExerciseSetDto(
     UUID id,
     Integer number,
-    Integer exerciseId,
-    Integer workoutId,
+    Long exerciseId,
+    Long workoutId,
     Integer reps,
     Float weight
 ) {}
